requirements.txt
asgiref==3.2.10
pytz==2020.1
sqlparse==0.2.4

Django>=2.2, <2.3
djangorestframework>=3.11, <4.0

python-decouple
celery>=4.4, <4.5
django-celery-beat
drf-yasg==1.17.1
sentry-sdk==0.14.2

psycopg2-binary
python-memcached
Pillow
zeep
coverage==5.1
mock==4.0.2
drf-compound-fields